// gateway/src/types/ENUMS.ts
export const enum E_STATUS {
  CREATE = 0,
  APPROVE = 1,
  DECLINE = 2,
  ARCHIVE = 3,
  DELAY = 4,
  REMOVE = 5,
  EMPTY_FILTER = 9,
}

export const enum E_SERVICE {
  ADMINISTRATOR = 0,
  VACANCY = 5,
}

export const enum E_ROLE {
  VIEW = 0,
  READ = 1,
  CREATE = 2,
  EDIT = 3,
  FULL = 4,
}

export enum E_SERVICES_CODE {
  profiles = '0084',
  news = '0002',
  vacancy = '0005',
}